% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.r
\name{aggregator}
\alias{aggregator}
\title{aggregator}
\usage{
aggregator(df = NULL, db = NULL, lat.field = "LATITUDE",
  lon.field = "LONGITUDE", rule.of = 5, agg.fields = NULL,
  sens.fields = NULL, agg.minutes = 5, quiet = FALSE)
}
\arguments{
\item{df}{a dataframe to be analyzed. If left \code{NULL}, a value for 
\code{db} should be provided}

\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values (in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude values (in decimal degrees)}

\item{rule.of}{default is \code{5} Whether or not a polygon is flagged as "CAN_SHOW" or not 
depends on the presence of a threshold number of unique values for certain sensitive fields.  
This parameter sets that threshold.}

\item{agg.fields}{the default is \code{NULL}, but if a db is provided, they 
will be guessed.  These are the fields on which to aggregate (i.e. generate 
values of \code{MEAN}, \code{COUNT} and \code{SUM}).}

\item{sens.fields}{the defaults are \code{NULL}  These are fields
to which the "rule of 5" should be applied. The Treasury Secretariat states that when data is 
shown to the public, certain fields must have at least 5 unique values for these fields 
aggregated together. When run, this function will look at these fields, and calculate how many 
unique values exist for each.  It will then populate a field 'PRIVUNIQUE' with the minimum number 
of unique values of all the assessed fields. If this is 5 or more, a field called 'CAN_SHOW' will 
be marked as 'YES' (otherwise it will be 'No').}

\item{agg.minutes}{the default is 5.  The number of minutes by which to aggregate.}

\item{quiet}{the default is \code{FALSE}.  If TRUE, no processing messages will be shown.}
}
\value{
a data frame
}
\description{
This function takes a dataframe and the number of minutes to aggregate by, and
returns the mean, count and sum for the specified field
}
\note{
This function is not a replacement for your brain.  If sensitive fields have names that are 
different than what is provided in the \code{sen.fields}, they will not be detected, or 
included in the checks.  Please make very sure you correctly identify such fields.
}
\seealso{
Other general_use: \code{\link{cleanup}},
  \code{\link{clip_by_poly}}, \code{\link{save_catches}},
  \code{\link{save_data}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
