% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_privacy.r
\name{assess_privacy}
\alias{assess_privacy}
\title{assess_privacy}
\usage{
assess_privacy(df = NULL, lat.field = "LATITUDE", lon.field = "LONGITUDE",
  rule.of = 5, agg.fields = NULL, sens.fields = NULL, nclasses = 4,
  show.plot = TRUE, save.plot = FALSE, fun = "mean",
  create.shps = FALSE, file_id = NULL, agg.poly.shp = NULL,
  agg.poly.field = NULL)
}
\arguments{
\item{df}{a dataframe to be analyzed. If left \code{NULL}, a value for 
\code{db} should be provided}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{rule.of}{default is \code{5} Whether or not data can be shown (even 
aggregated) depends on the presence of a threshold number of unique values 
for certain sensitive fields.  This parameter sets that threshold.}

\item{agg.fields}{the default is \code{NULL}, but if a db is provided, they 
will be guessed.  These are the fields on which to aggregate (i.e. generate 
values of \code{MEAN}, \code{COUNT} and \code{SUM}).}

\item{sens.fields}{the defaults are \code{NULL}  These are fields
to which the "rule of 5" should be applied. The Treasury Secretariat states that when data is 
shown to the public, certain fields must have at least 5 unique values for these fields 
aggregated together. When run, this function will look at these fields, and calculate how many 
unique values exist for each.  It will then populate a field 'TOTUNIQUE' with the minimum number 
of unique values of all the assessed fields. If this is 5 or more, a field called 'CAN_SHOW' will 
be marked as 'YES' (otherwise it will be 'NO').}

\item{nclasses}{default is \code{4} This is how many discrete classes you 
want to use to categorize data on the output plot.}

\item{show.plot}{default is \code{TRUE} This indicates whether or not a plot 
of your aggregated data should be shown.}

\item{save.plot}{default is \code{FALSE} This indicates whether or not you 
would like a png file saved of your data. If TRUE - a plot will be saved 
INSTEAD of being shown}

\item{fun}{the default is \code{mean}.  This is the aggregated data that 
will be used to symnbolize your plotted data.  Options include "mean", "sum" 
and "length" (which is a count)}

\item{create.shps}{default is \code{FALSE}.  This indicates whether or not 
shapefiles should be created for 1) the polygon file (with aggregated values 
for each polygon and an indication of whether or not each polygon meets the 
privacy constraints), and 2) the 2 min gridded data (only for within those 
polygons that meet the privacy constraints).}

\item{file_id}{default is \code{NULL} Whatever is entered here will be used 
to name the output shapefiles and/or plots.  If nothing is enetered, the 
output files will just be named using timestamps.}

\item{agg.poly.shp}{default is \code{NULL}.  This is the shapefile that has 
polygons that should be checked for sufficient unique values of the 
sens.fields.  If NULL, NAFO zones will be used.  Otherwise, a path to any 
polygon shapefile can be provided.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within 
the shapefile provided to agg.poly.shp that should be used to check for 
sufficient unique values of the sens.fields.}
}
\value{
a SpatialPolygonsDataFrame, and generates a shapefile
}
\description{
At this time, data with privacy considerations must be aggregated such that each 
polygon has a minimum of 5 unique values for sensitive fields like Licenses, License Holders, and 
Vessels.  This function takes a dataframe and shapefile and for each polygon in the 
shapefile calculates 1) aggregate values for a number of (user-specified) fields , and 2) 
how many unique values exist in each polygon for each of a number of sensitive fields. A 
shapefile is generated with all of the data, as well as a field indicating whether or not the 
data can be displayed.
}
\note{
If sensitive fields have names that are 
different than what is provided in the \code{sen.fields}, they will not be detected, or 
included in the checks.  Please make very sure you correctly identify such fields.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
