% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.r
\name{cleanup}
\alias{cleanup}
\title{cleanup}
\usage{
cleanup(db = NULL, var.like = NULL)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{var.like}{The default value is NULL.  Providing a value to this parameter will remove ALL
elements from your environment that match it.  Be careful and be as specific as you can.}
}
\description{
This function removes the R objects associated with the various objects created
by each of the databases it connects to.
}
\seealso{
Other general_use: \code{\link{clip_by_poly}},
  \code{\link{get_isdbTripSet}},
  \code{\link{save_catches}}, \code{\link{save_data}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
