% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tweaks.r
\name{data_tweaks}
\alias{data_tweaks}
\title{data_tweaks}
\usage{
data_tweaks(db = NULL, data.dir = file.path(getwd(), "data"))
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{data.dir}{The default is a "data" folder within your working directory. This is the path to where you want the
extracted files to go.}
}
\description{
After the initial extraction, this function is called to do minor housekeeping
to the extracted datasets to ensure they works with the other functions.
}
\note{
data is not added to the global environment by this function - changes are made, saved, and
dropped.
the following are special data handling processes specific to the ISDB tables
the following are special data handling processes specific to the rv tables (beyond
getting the whole table)
the following are special data handling processes specific to the rvp70 tables (beyond
getting the whole table)
}
\seealso{
Other internal: 
\code{\link{load_datasources}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
