% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tweaks2.R
\name{data_tweaks2}
\alias{data_tweaks2}
\title{data_tweaks2}
\usage{
data_tweaks2(db = "ALL", extract_user = NULL, extract_computer = NULL)
}
\arguments{
\item{db}{default is \code{"ALL"}. This identifies the dataset you are working 
with.}

\item{extract_user}{default is \code{NULL}.  This parameter can be used with
\code{extract_computer} to load encypted data files extracted by another user 
and/or computer}

\item{extract_computer}{default is \code{NULL}.  This parameter can be used with
\code{extract_user} to load encypted data files extracted by another user 
and/or computer}
}
\description{
After the initial extraction, this function is called to do minor housekeeping
to the extracted datasets to ensure they works with the other functions.
}
\note{
data is not added to the global environment by this function - changes are made, saved, and
dropped.
the following are special data handling processes specific to the GROUNDFISH tables (beyond
getting the whole table)
}
\seealso{
Other internal: 
\code{\link{load_datasources}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
