% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\docType{data}
\name{encrypted_schemas}
\alias{encrypted_schemas}
\title{Global configuration for encrypted schemas}
\format{
An object of class \code{character} of length 4.
}
\usage{
encrypted_schemas
}
\description{
Global configuration for encrypted schemas
}
\keyword{internal}
