% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fleet.r
\name{get_fleet}
\alias{get_fleet}
\title{get_fleet}
\usage{
get_fleet(fn.oracle.username = "_none_", fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_", usepkg = "rodbc", dateStart = NULL,
  dateEnd = NULL, mdCode = NULL, sectors = 7, data.dir = getwd())
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for this stored
in your environment (e.g. from an rprofile file), this can be left and that
value will be used.  If a value for this is provided, it will take priority
over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of 
data will be returned.}

\item{mdCode}{default is \code{NULL}. This is the MARFIS Monitoring
Document code used to identify fleets. If this is left as NULL, a popup
will allow the user to select one from a list.
Valid codes include numbers 1-53, including
\itemize{
\item 2 = "MOBILE GEAR GROUNDFISH"
\item 5 = "SWORDFISH/SHARK LONGLINE"
\item 10 = "CRAB"
\item 12 = "INSHORE DREDGE SHELLFISH"
\item 13 = "OFFSHORE CLAM"
\item 14 = "SCALLOP"
\item 15 = "OFFSHORE SCALLOP"
\item 16 = "MOBILE SHRIMP"
\item 18 = "SEA URCHIN"
\item 19 = "OFFSHORE LOBSTER"
\item 20 = "FIXED GEAR (GROUNDFISH)"
\item 21 = "GROUNDFISH - SHRIMP"
\item 22 = "ATLANTIC BLUEFIN TUNA"
\item 25 = "CRAB LOG BOOK"
\item 23 = "PELAGICS"
\item 28 = "SHRIMP TRAP"
\item 31 = "NORTHERN SHRIMP SLIP"
\item 39 = "SEA CUCUMBER"
\item 41 = "SEA CUCUMBER (2012)"
\item 47 = "HAGFISH"
\item 49 = "LOBSTER 38 B"
\item 52 = "SEA URCHIN FOR LFA 36 & 38"
}}

\item{sectors}{default is \code{7}. This identifies the region.}

\item{data.dir}{The default is your working directory. If you are hoping to 
load existing data, this folder should identify the folder containing your 
*.rdata files.}
}
\value{
returns a data.frame with 4 columns - "MON_DOC_DEFN_ID", "MON_DOC_ID", 
"LICENCE_ID" and "VR_NUMBER".
}
\description{
This function extracts all of the Vessels, Licences and MARFIS 
Monitoring Doc IDs associated with a particular fleet for a particular date 
range.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
