% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isdbTripSet.R
\name{get_isdbTripSet}
\alias{get_isdbTripSet}
\title{get_isdbTripSet}
\usage{
get_isdbTripSet(FS_ID = NULL)
}
\arguments{
\item{FS_ID}{default is \code{NULL}. This is one or more FISHSET_IDs for 
which you want the associated trip name and set number.}
}
\value{
DataFrame
}
\description{
This function takes  one or more FISHSET_IDs and returns the 
name and set number for the FISHSET_ID.
}
\seealso{
Other general_use: \code{\link{cleanup}},
  \code{\link{clip_by_poly}}, \code{\link{save_catches}},
  \code{\link{save_data}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
