% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_rv.r
\name{get_std_rv}
\alias{get_std_rv}
\title{get_std_rv}
\usage{
get_std_rv(fn.oracle.username = "_none_",
  fn.oracle.password = "_none_", fn.oracle.dsn = "_none_",
  usepkg = "rodbc", data.dir = NULL, year_min = NULL,
  year_max = NULL, season = NULL, strat = NULL, sp = NULL,
  len_min = NULL, len_max = NULL)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{data.dir}{The default is your working directory. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. If you are extracting data,
a data folder will be created under this folder.
extracted files to go.}

\item{year_min}{default is \code{NULL}.  This is the minimum year you want data to be returned for.}

\item{year_max}{default is \code{NULL}. This is the maximum year you want data to be returned for.}

\item{season}{default is \code{NULL}. This is the season you want data for (e.g. "SUMMER")}

\item{strat}{default is \code{NULL}.  These are the strata you want results for (e.g. \code{c(440:495)})}

\item{sp}{default is \code{NULL}.  This is the species code you want data for}

\item{len_min}{default is \code{NULL}.  This is the minimum length you want data to be returned for.}

\item{len_max}{default is \code{NULL}. This is the maximum length you want data to be returned for.}
}
\description{
This function prepares a detailed length frequency table based on 1, 2 or 3 cm intervals
depending on the species being processed. It provides the same data as results from NWAGS.gsd3lf_mv.
}
\seealso{
Other dfo_extractions: \code{\link{data_filter}},
  \code{\link{get_data_custom}}, \code{\link{get_data}},
  \code{\link{self_filter}},
  \code{\link{summarize_catches}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
