% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_survey}
\alias{get_survey}
\title{get_survey}
\usage{
get_survey(
  db = NULL,
  survey = NULL,
  keepBadSets = FALSE,
  data.dir = NULL,
  quiet = TRUE,
  env = .GlobalEnv
)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working
with.}

\item{survey}{default is \code{NULL}. This identifies which survey you want tto extract. Valid 
values include the following:
\itemize{
\item \code{4X} - Type 1; Spring (i.e. months 1:4); 2008+; specific strata 
\item \code{GEORGES} - Type 1; Spring (i.e. months 1:4); strata 5Z*
\item \code{SPRING} - Type 1; Spring (i.e. months 1:4); pre-2008; specific strata 
\item \code{4VSW}  - Type 1; Spring (i.e. months 1:4); 4VSW strata;  
\item \code{SUMMER} - Type 1; Summer (i.e. months 5:8); specific strata
\item \code{FALL} - Type 1; Fall (i.e. months 9:12)
}}

\item{keepBadSets}{default is \code{FALSE}.  This determines whether or not both type 1 and 3 
sets are returned, or just type 1. Type 3 sets indicate an invalid tow.}

\item{data.dir}{The default is NULL. This should be a path to a folder containing your rdata files.}

\item{quiet}{default is \code{TRUE}.  If TRUE, no output messages will be shown.}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}
}
\description{
This function facilitates the extraction of particular survey from the rv database.
Within the groundfish data are a number of survey all mixed together.  They are typically 
differentiated via specific selections of 1) type; 2) date (i.e. month and year), and 3) strata.
Since this can be confusing, this function was written to help ensure that the extracted sets 
match the survey they should.
}
\seealso{
Other dfo_extractions: 
\code{\link{get_data}()},
\code{\link{get_data_custom}()},
\code{\link{self_filter}()},
\code{\link{summarize_catches}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
