% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giddyup.r
\name{giddyup}
\alias{giddyup}
\title{giddyup}
\usage{
giddyup(db = NULL, usepkg = NULL, force.extract = FALSE, QC = FALSE)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{usepkg}{default is \code{'rodbc'}. his indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{force.extract}{The default value is FALSE.  By default, existing data will be loaded.  If
\code{force.extract ==TRUE}, than a full extraction will take place, overwriting any existing
data.}

\item{QC}{The default value is FALSE.  In most large datasets are 'orphan' records that can't be
joined to other records.  These are normally ignored and discarded. If  \code{QC == TRUE}, they
will be retained so that they can be investigated.
\code{force.extract ==TRUE}, than a full extraction will take place, overwriting any existing
data.}
}
\description{
This function walks people through an extraction, by helping select a database, and
and R/Oracle connection package.
}
\note{
When prompted for a way to connect to Oracle, most people will select 'RODBC' since it's
easier to install, but ROracle is MUCH faster.
}
\seealso{
Other dfo_extractions: \code{\link{data_filter}},
  \code{\link{get_data}}, \code{\link{make_log}},
  \code{\link{make_oracle_cxn}}, \code{\link{self_filter}},
  \code{\link{summarize_catches}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
