% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc_grid2Min.r
\docType{data}
\name{grid2Min}
\alias{grid2Min}
\title{This is the contents of the grid2Min rdata file}
\description{
This is the contents of the grid2Min rdata file
}
\references{
\url{data_blah.com}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\keyword{data}
