% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasources.r
\name{load_datasources}
\alias{load_datasources}
\title{load_datasources}
\usage{
load_datasources(db = NULL)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}
}
\description{
This file was introduced in version 2.0 of Mar.datawrangling as an attempt simplify the process
of adding additional data sources.
If done correctly, this should remove the need for a variety of files and 
generally clarify the code.
}
\seealso{
Other internal: \code{\link{data_tweaks}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
