% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_log.r
\name{make_log}
\alias{make_log}
\title{make_log}
\usage{
make_log(refresh.data = FALSE)
}
\arguments{
\item{refresh.data}{default is \code{FALSE}.  If TRUE, the existing activity_
log will be erased and replaced with a new one.}
}
\value{
a list of lists describing what filters have been applied to the data
}
\description{
This function tracks the users' filtering selections by writing 
them to an object called "activity_log".  It tracks when the data was 
loaded, as well as what filters were applied.
}
\seealso{
Other dfo_extractions: \code{\link{data_filter}},
  \code{\link{get_data}}, \code{\link{giddyup}},
  \code{\link{make_oracle_cxn}}, \code{\link{self_filter}},
  \code{\link{summarize_catches}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
