% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_oracle_cxn.r
\name{make_oracle_cxn}
\alias{make_oracle_cxn}
\title{make_oracle_cxn}
\usage{
make_oracle_cxn(usepkg = "rodbc", fn.oracle.username = "_none_",
  fn.oracle.password = "_none_", fn.oracle.dsn = "_none_")
}
\arguments{
\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for this stored
in your environment (e.g. from an rprofile file), this can be left and that
value will be used.  If a value for this is provided, it will take priority
over your existing value.}
}
\description{
This function facilitates creating a connection to Oracle, and
allows connection via RODBC or ROracle, depending on the value of usepkg.
}
\seealso{
Other dfo_extractions: \code{\link{data_filter}},
  \code{\link{get_data}}, \code{\link{make_log}},
  \code{\link{self_filter}},
  \code{\link{summarize_catches}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
