% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_catches.r
\name{save_catches}
\alias{save_catches}
\title{save_catches}
\usage{
save_catches(db = NULL, df = NULL, filename = NULL,
  df.crs = "+init=epsg:4326", req.coords = TRUE, lat.field = "LATITUDE",
  lon.field = "LONGITUDE", keep_nullsets = FALSE, formats = c("csv",
  "shp"))
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working
with.}

\item{df}{This is a dataframe that you want to save in some other format.  If a spatial format is selected
(e.g. shapefile), it must have LATITUDE and LONGITUDE fields}

\item{filename}{default is \code{NULL}.  This will be the prefix of your filename}

\item{df.crs}{This is the CRS value for your dataframe.  This should be the reference system that your data is known to be in.
The default value \code{"+init=epsg:4326"} is WGS84 and is appropriate for most data collected using a GPS.}

\item{req.coords}{default is TRUE.  This filters out records without values for LATITUDE or
LONGITUDE.  The function aborts if req.coords=TRUE and no records remain.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values (in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude values (in decimal degrees)}

\item{keep_nullsets}{default is FALSE.  In many cases, in addition to the
catch, it is also useful to know where fishing occurred that did not result
in catches.  this is especially true for most survey-type data.
Note that for industry fishing (i.e. ISDB and MARFIS), the null sets will
actually just consist of all sets that met your filter criteria.}

\item{formats}{This is a vector of the formats in which you would like to save the current data}
}
\description{
This function takes all of the data you've filtered, and rolls it up into a csv and/
or a shapefile for continued analysis
}
\seealso{
Other general_use: \code{\link{cleanup}},
  \code{\link{clip_by_poly}},
  \code{\link{get_isdbTripSet}}, \code{\link{save_data}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
