% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tables.R
\name{save_tables}
\alias{save_tables}
\title{save_tables}
\usage{
save_tables(db = NULL)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}
}
\value{
nothing
}
\description{
This function takes a snapshot of all of the tables of a database as they currently
exist in your environment, and stores them in a new environment (i.e.'dw').
}
\seealso{
Other general_use: 
\code{\link{cleanup}()},
\code{\link{clip_by_poly}()},
\code{\link{get_isdb_trip_set}()},
\code{\link{restore_tables}()},
\code{\link{save_data}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
