% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_filter.r
\name{self_filter}
\alias{self_filter}
\title{self_filter}
\usage{
self_filter(db = NULL, looponce = FALSE, debug = FALSE)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{looponce}{default is \code{NULL}.  This is only used for QC purposes.}

\item{debug}{default is \code{FALSE}.  This is used for debugging, and will
cause the filtering commands to be printed out.}
}
\description{
This function draws upon the relationships between the tables 
reported in load_datasources() so that if one is filtered via the GUI, or via
a subset like \code{ISGEARCODES = ISGEARCODES[ISGEARCODES$GEARCD_ID==7,]}), 
all other tables will be filtered until the remaining records all relate to 
eachother.
For example, if the ISVESSELS table was filtered to only contain a single 
vessel - all other tables would be filtered until all remaing records in all 
tables relate directly to that vessel.
}
\note{
This line is here to prevent an error message
Count the rows in all of the tables for this dataset
}
\seealso{
Other dfo_extractions: \code{\link{data_filter}},
  \code{\link{get_data}}, \code{\link{giddyup}},
  \code{\link{make_log}}, \code{\link{make_oracle_cxn}},
  \code{\link{summarize_catches}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
