% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.r
\name{cleanup}
\alias{cleanup}
\title{cleanup}
\usage{
cleanup(db = NULL, var.like = NULL, env = .GlobalEnv)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working 
with.}

\item{var.like}{The default value is NULL.  Providing a value to this parameter will remove ALL
elements from your environment that match it.  Be careful and be as specific as you can.}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}
}
\description{
This function removes the R objects associated with the various objects created
by each of the databases it connects to.
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly}()},
\code{\link{get_isdb_trip_set}()},
\code{\link{restore_tables}()},
\code{\link{save_data}()},
\code{\link{save_tables}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
