% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tweaks2.R
\name{data_tweaks2}
\alias{data_tweaks2}
\title{data_tweaks2}
\usage{
data_tweaks2(db = "ALL")
}
\arguments{
\item{db}{default is \code{"ALL"}. This identifies the dataset you are working 
with.}
}
\description{
After the initial extraction, this function is called to do minor housekeeping
to the extracted datasets to ensure they works with the other functions.
}
\note{
data is not added to the global environment by this function - changes are made, saved, and
dropped.
the following are special data handling processes specific to the GROUNDFISH tables (beyond
getting the whole table)
}
\seealso{
Other internal: 
\code{\link{load_datasources}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
