% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_custom.r
\name{get_data_custom}
\alias{get_data_custom}
\title{get_data_custom}
\usage{
get_data_custom(
  schema = NULL,
  data.dir = file.path(getwd(), "data"),
  tables = NULL,
  usepkg = "rodbc",
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  env = .GlobalEnv,
  quiet = F
)
}
\arguments{
\item{schema}{default is \code{NULL}. This is the schema you want to access 
a additional tables from.}

\item{data.dir}{The default is your working directory. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. If you are extracting data,
a data folder will be created under this folder.
extracted files to go.}

\item{tables}{The default value is \code{NULL}.  This is a vector of table 
names you want to extract that exist in the database specified by \code{db}.}

\item{usepkg}{default is \code{oracle_cxn$usepkg}. This indicates whether the 
connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to 
connect.  rodbc is slightly easier to setup, but roracle will extract data ~ 
5x faster.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}

\item{quiet}{default is \code{FALSE}.  If TRUE, no output messages will be shown.}
}
\description{
This function allows users to extract other tables within a
known database that the script does not get by default.  They must have 
access to the tables to extract them.
}
\seealso{
Other dfo_extractions: 
\code{\link{get_data}()},
\code{\link{get_survey}()},
\code{\link{self_filter}()},
\code{\link{summarize_catches}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
