% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc_NAFOSubunits.r
\docType{data}
\name{NAFOSubunits}
\alias{NAFOSubunits}
\title{This is the contents of the NAFOSubunits rdata file}
\description{
This is the contents of the NAFOSubunits rdata file
}
\references{
\url{data_blah.com}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\keyword{data}
