% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_findorphans.r
\name{qc_findorphans}
\alias{qc_findorphans}
\title{qc_findorphans}
\usage{
qc_findorphans(db = NULL, debug = F)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working
with.}

\item{debug}{default = \code{FALSE}.  This is passed to self_filter() which 
causes each filter to be printed out.}
}
\description{
This function creates orph_* tables with records that don't
appear to be linked to the other tables (within Mar.datawrangling)
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
