% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_filter.r
\name{self_filter}
\alias{self_filter}
\title{self_filter}
\usage{
self_filter(
  db = NULL,
  looponce = FALSE,
  debug = FALSE,
  env = .GlobalEnv,
  keep_nullsets = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working
with.}

\item{looponce}{default is \code{NULL}.  This is only used for QC purposes.}

\item{debug}{default is \code{FALSE}.  This is used for debugging, and will
cause the filtering commands to be printed out.}

\item{env}{This the the environment you want this function to work in.  The
default value is \code{.GlobalEnv}.}

\item{keep_nullsets}{default is \code{TRUE}  If you're working the data that can
be extracted via this package, it can be useful to know where fishing
occurred that did not result in catches.  Note that for industry fishing
(i.e. ISDB, MARFIS, COMLAND**), this will just ensure that all sets matching
your criteria are retained.  If you have selected data by the "caught
species", you will have already exluded the nullsets.}

\item{quiet}{default is \code{FALSE}.  If True, the filtering process will
not show text describing progress.}
}
\description{
This function draws upon the relationships between the tables
reported in load_datasources() so that if one is filtered via the GUI, or via
a subset like \code{ISGEARCODES = ISGEARCODES[ISGEARCODES$GEARCD_ID==7,]}),
all other tables will be filtered until the remaining records all relate to
eachother.
For example, if the ISVESSELS table was filtered to only contain a single
vessel - all other tables would be filtered until all remaing records in all
tables relate directly to that vessel.
}
\note{
This line is here to prevent an error message
}
\seealso{
Other dfo_extractions: 
\code{\link{get_data}()},
\code{\link{get_data_custom}()},
\code{\link{get_survey}()},
\code{\link{summarize_catches}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
