% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_fishing_picture.r
\name{total_fishing_picture}
\alias{total_fishing_picture}
\title{total_fishing_picture}
\usage{
total_fishing_picture(fn.oracle.username = "_none_",
  fn.oracle.password = "_none_", fn.oracle.dsn = "_none_",
  usepkg = "rodbc", dateStart = NULL, dateEnd = NULL,
  mdCode = NULL, maxBreak_mins = 1440, obsCovByArea = FALSE,
  agg.poly.shp = "NAFO", agg.poly.field = "NAFO_1", data.dir = NULL,
  qplot = FALSE, quiet = FALSE)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of 
data will be returned.}

\item{mdCode}{default is \code{NULL}. This is the MARFIS Monitoring
Document code used to identify fleets. If this is left as NULL, a popup
will allow the user to select one from a list.}

\item{maxBreak_mins}{default is \code{1440} (minutes, or 24 hours).  This is the 
amount of time that can pass for a vessel between successive VMS points before 
a new "trek" is imposed.  For example, if a vessel sits motionless at port 
for 13 hours before heading out again, the new data will count towards the 
previous "trek".  However, if it sits for 25 hours, a new trek will be created.
This does not change the data that is extracted, just how the data is broken 
into discrete line segments.}

\item{obsCovByArea}{default is \code{FALSE}. Observer coverage is calculated 
simply as a percentage of observer trips vs all trips.  If this is set to TRUE,
coverage will be calculated on an area by area basis using the shapefile identified 
in \code{agg_poly_shape} and the field identified in \code{agg.poly.field}.}

\item{agg.poly.shp}{default is \code{NULL}.  This is the shapefile that has 
polygons that should be checked for sufficient unique values of the 
sens.fields.  If NULL, NAFO zones will be used.  Otherwise, a path to any 
polygon shapefile can be provided.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within 
the shapefile provided to agg.poly.shp that should be used to check for 
sufficient unique values of the sens.fields.}

\item{data.dir}{The default is your working directory. If you are hoping to 
load existing data, this folder should identify the folder containing your 
*.rdata files.}

\item{qplot}{default is \code{FALSE}. If set to TRUE, a quick and dirty plot 
of the vms (grey (unobserved) and blue (observed)), Marfis (small black crosses) 
and Observer data (blue dots) will be generated.}

\item{quiet}{default is \code{FALSE}. If set to TRUE, no output messages will
be displayed as the script runs.}
}
\value{
a list of 6 objects - "obs_raw" & "marfis_raw" contain the data extracted
from their respected databases.  "obs_sp" & "marfis_sp" are the same data, but as 
SpatialPointsDataFrames. "vmstracks" is a SpatialLinesDataFrame with identified 
"treks". "obs_coverage" is a table showing the number of observed trips vs the 
total number of trips (by area if obsCovByArea is TRUE)
}
\description{
This function extracts all of the observer, Marfis and VMS data 
for a particular date range, for a particular Fleet (identified by the 
monitoring document code).
}
\seealso{
Other fleets: \code{\link{get_fleet}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
